Move game-specific watcher code in with the game setup code.
